// WYNKWeb/src/app/Models/over-time.ts
export class OverTime {
  CMPID: number;
  EmpID: number;
  FromTime: string;
  ToTime: string;

  //FromTime: Date;
  //ToTime: Date;
  NoofHours: number;
  OTperHour: number;
  TotalOTAmount: number;
  OTDate: Date;
  IsProcessed: boolean;
  ProcessedDate: Date;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;

}